bundle scripts: don't update repo if -h option is invoked

+++ Bundlebot/smv/scripts/run_smvbundle.sh
@@ -1,8 +1,11 @@
 #!/bin/bash
+opt=$1
 
-echo "*** updating bot repo"
-git clean -dxf          >& /dev/null
-git remote update       >& /dev/null
-git merge origin/master >& /dev/null
+if [ "$opt" != "-h" ]; then
+  echo "*** updating bot repo"
+  git clean -dxf          >& /dev/null
+  git remote update       >& /dev/null
+  git merge origin/master >& /dev/null
+fi
 ./smvbundle.sh $*
